Vision/pyproject.toml
[tool.poetry]
name = "FRC Vision Practice"
version = "0.1.0"
description = ""
authors = ["Will <williamhaley@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.5.5"
pynetworktables = "^2021.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vision = "vision:main"
